ST-421: Move TaxCodechange to new file
