+++ app/uk/gov/hmrc/tai/model/api/TaxCodeSummary.scala
@@ -20,7 +20,6 @@ import org.joda.time.LocalDate
 import play.api.libs.json.{Json, OFormat}
 import uk.gov.hmrc.tai.model.TaxCodeRecord
 import uk.gov.hmrc.tai.model.tai.TaxYear
-import uk.gov.hmrc.tai.util.DateTimeHelper.dateTimeOrdering
 import uk.gov.hmrc.time.TaxYearResolver
 
 case class TaxCodeSummary(taxCode: String,
@@ -53,32 +52,4 @@ object TaxCodeSummary {
     )
   }
 
-}
-
-case class TaxCodeChange(current: Seq[TaxCodeSummary], previous: Seq[TaxCodeSummary]) {
-  def latestTaxCodeChangeDate: LocalDate = current.map(_.startDate).min
-
-  def primaryCurrentTaxCode: String = primaryTaxCode(current)
-  def secondaryCurrentTaxCodes: Seq[String] = secondaryTaxCode(current)
-  def primaryPreviousTaxCode: String = primaryTaxCode(previous)
-  def secondaryPreviousTaxCodes: Seq[String] = secondaryTaxCode(previous)
-
-  def primaryCurrentPayrollNumber: Option[String] = primaryPayrollNumber(current)
-  def secondaryCurrentPayrollNumbers: Seq[String] = secondaryPayrollNumbers(current)
-  def primaryPreviousPayrollNumber: Option[String] = primaryPayrollNumber(previous)
-  def secondaryPreviousPayrollNumbers: Seq[String] = secondaryPayrollNumbers(previous)
-
-  private def primaryPayrollNumber(records: Seq[TaxCodeSummary]) = primaryRecord(records).payrollNumber
-  private def secondaryPayrollNumbers(records: Seq[TaxCodeSummary]) = secondaryRecords(records).flatMap(_.payrollNumber)
-  private def primaryTaxCode(records: Seq[TaxCodeSummary]) = primaryRecord(records).taxCode
-  private def secondaryTaxCode(records: Seq[TaxCodeSummary]) = secondaryRecords(records).map(_.taxCode)
-
-  private def primaryRecord(records: Seq[TaxCodeSummary]) =
-    records.find(_.primary).getOrElse(throw new RuntimeException("No primary tax code record found"))
-
-  private def secondaryRecords(records: Seq[TaxCodeSummary]) = records.filterNot(_.primary)
-}
-
-object TaxCodeChange {
-  implicit val format: OFormat[TaxCodeChange] = Json.format[TaxCodeChange]
 }
